+++ st2api/st2api/app.py
@@ -55,7 +55,7 @@ def setup_app(config=None):
 
     app_conf = dict(config.app)
 
-    active_hooks = [hooks.JSONErrorResponseHook(), hooks.LoggingHook()]
+    active_hooks = [hooks.RequestIDHook(), hooks.JSONErrorResponseHook(), hooks.LoggingHook()]
 
     if cfg.CONF.auth.enable:
         active_hooks.append(hooks.AuthHook())