+++ st2api/st2api/controllers/v1/actionexecutions.py
@@ -29,6 +29,7 @@ from st2api.controllers.v1.executionviews import SUPPORTED_FILTERS
 from st2common import log as logging
 from st2common.constants.action import LIVEACTION_STATUS_CANCELED
 from st2common.constants.action import CANCELABLE_STATES
+from st2common.exceptions.trace import TraceNotFoundException
 from st2common.models.api.action import LiveActionAPI
 from st2common.models.api.base import jsexpose
 from st2common.models.api.execution import ActionExecutionAPI
@@ -113,6 +114,8 @@ class ActionExecutionsControllerMixin(BaseRestControllerMixin):
         except jsonschema.ValidationError as e:
             LOG.exception('Unable to execute action. Parameter validation failed.')
             abort(http_client.BAD_REQUEST, re.sub("u'([^']*)'", r"'\1'", e.message))
+        except TraceNotFoundException as e:
+            abort(http_client.BAD_REQUEST, str(e))
         except Exception as e:
             LOG.exception('Unable to execute action. Unexpected error encountered.')
             abort(http_client.INTERNAL_SERVER_ERROR, str(e))