+++ st2client/st2client/models/__init__.py
@@ -20,4 +20,5 @@ from st2client.models.action_alias import *  # noqa
 from st2client.models.keyvalue import *    # noqa
 from st2client.models.policy import *       # noqa
 from st2client.models.reactor import *      # noqa
+from st2client.models.trace import *      # noqa
 from st2client.models.webhook import *      # noqa