+++ st2common/st2common/models/api/sensor.py
@@ -43,6 +43,9 @@ class SensorTypeAPI(BaseAPI):
             'description': {
                 'type': 'string'
             },
+            'enabled': {
+                'type': 'boolean',
+            },
             'artifact_uri': {
                 'type': 'string',
             },