+++ st2reactor/st2reactor/rules/worker.py
@@ -16,7 +16,9 @@
 from kombu import Connection
 
 from st2common import log as logging
+from st2common.constants.trace import TRACE_CONTEXT, TRACE_ID
 from st2common.util import date as date_utils
+from st2common.services.trace import add_or_update_given_trace_context
 from st2common.transport import consumers, reactor
 from st2common.transport import utils as transport_utils
 import st2reactor.container.utils as container_utils
@@ -55,6 +57,16 @@ class TriggerInstanceDispatcher(consumers.MessageHandler):
 
         if trigger_instance:
             try:
+                # Use trace_context from the instance and if not found create a new context
+                # and use the trigger_instance.id as trace_tag.
+                trace_context = instance.get(TRACE_CONTEXT, None)
+                if not trace_context:
+                    trace_context = {
+                        TRACE_ID: 'trigger_instance-%s' % str(trigger_instance.id)
+                    }
+                # add a trace or update an existing trace with trigger_instance
+                add_or_update_given_trace_context(trace_context=trace_context,
+                                                  trigger_instances=[str(trigger_instance.id)])
                 self.rules_engine.handle_trigger_instance(trigger_instance)
             except:
                 # This could be a large message but at least in case of an exception