+++ st2tests/st2tests/config.py
@@ -162,10 +162,13 @@ def _register_action_sensor_opts():
 
 def _register_mistral_opts():
     mistral_opts = [
-        cfg.StrOpt('v2_base_url', default='http://localhost:8989/v2',
-                   help='Mistral v2 API server root endpoint.'),
-        cfg.IntOpt('max_attempts', default=2),
-        cfg.IntOpt('retry_wait', default=1)
+        cfg.StrOpt('v2_base_url', default='http://localhost:8989/v2', help='v2 API root endpoint.'),
+        cfg.IntOpt('max_attempts', default=2, help='Max attempts to reconnect.'),
+        cfg.IntOpt('retry_wait', default=1, help='Seconds to wait before reconnecting.'),
+        cfg.StrOpt('keystone_username', default=None, help='Username for authentication.'),
+        cfg.StrOpt('keystone_password', default=None, help='Password for authentication.'),
+        cfg.StrOpt('keystone_project_name', default=None, help='OpenStack project scope.'),
+        cfg.StrOpt('keystone_auth_url', default=None, help='Auth endpoint for Keystone.')
     ]
     _register_opts(mistral_opts, group='mistral')